ai_news_generator/requirements.txt
crewai
crewai-tools
streamlit
python-dotenv
protobuf>=3.20.3,<5.0.0
